+++ src/features/problems/queries/index.ts
@@ -2,4 +2,3 @@ export * from "./getProblem";
 export * from "./getProblemScore";
 export * from "./getProblemsScore";
 export * from "./recalcProblemScore";
-export * from "./submit";
