+++ src/features/submissions/queries/index.ts
@@ -1,1 +1,3 @@
+export * from "./calcSubmissionScore";
 export * from "./getSubmissionDiff";
+export * from "./submit";
